Add task number as task parameter
That way tasks have a reference to "self" and can modify the tasks from within.
e.g. create one-shots, multi-shots, or tasks with varying intervals.

This commit comes including changes I made to the files for some other project and are squashed with changes to the naming throughout the lib and therefore breaks compatibility.